requirements.txt
altgraph==0.17.2
cmake==3.22.2
et-xmlfile==1.1.0
future==0.18.2
lxml==4.7.1
numpy==1.22.2
openpyxl==3.0.9
pandas==1.5.1
pefile==2021.9.3
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pywin32-ctypes==0.2.0
six==1.16.0
xlrd==2.0.1
